pmemcheck: add small_stacktrace test suite
Multi-threaded tests on some OSes do not print the stacktrace from
pthread. Remove excess stacktrace.
