+++ pmemcheck/tests/small_stacktrace/trans_mt.vgtest
