+++ pmemcheck/tests/small_stacktrace/trans_mt_cross.vgtest
